Compile *.proto files to gRPC
In order to compile retrieved elements from containerd and its dependency (gogoproto), I had to replace several "import" statements.
I prefer to do that with string replacement, because I have not found how to do that with configuration of gradle-protobuf-plugin.
Add explicit dependencies between tasks to insure correctness of build.